+++ launchdarkly-android-client/src/main/java/com/launchdarkly/android/ConnectivityReceiver.java
@@ -3,17 +3,17 @@ package com.launchdarkly.android;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.util.Log;
+
+import timber.log.Timber;
 
 import static com.launchdarkly.android.Util.isInternetConnected;
 
 public class ConnectivityReceiver extends BroadcastReceiver {
-    private static final String TAG = "LDConnectivityReceiver";
 
     @Override
     public void onReceive(Context context, Intent intent) {
         if (isInternetConnected(context)) {
-            Log.d(TAG, "Connected to the internet");
+            Timber.d("Connected to the internet");
             try {
                 LDClient ldClient = LDClient.get();
                 if (!ldClient.isOffline()) {
@@ -24,15 +24,15 @@ public class ConnectivityReceiver extends BroadcastReceiver {
                     }
                 }
             } catch (LaunchDarklyException e) {
-                Log.e(TAG, "Tried to restart foreground updating, but LDClient has not yet been initialized.");
+                Timber.e(e, "Tried to restart foreground updating, but LDClient has not yet been initialized.");
             }
         } else {
-            Log.d(TAG, "Not Connected to the internet");
+            Timber.d("Not Connected to the internet");
             try {
                 LDClient ldClient = LDClient.get();
                 ldClient.stopForegroundUpdating();
             } catch (LaunchDarklyException e) {
-                Log.e(TAG, "Tried to stop foreground updating, but LDClient has not yet been initialized.");
+                Timber.e(e, "Tried to stop foreground updating, but LDClient has not yet been initialized.");
             }
         }
     }