+++ launchdarkly-android-client/src/main/java/com/launchdarkly/android/LDClient.java
@@ -4,7 +4,6 @@ import android.app.Application;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.support.annotation.NonNull;
-import android.util.Log;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
@@ -20,6 +19,7 @@ import com.google.gson.JsonSyntaxException;
 
 import java.io.Closeable;
 import java.io.IOException;
+import java.lang.ref.WeakReference;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.Callable;
@@ -29,6 +29,8 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import timber.log.Timber;
+
 import static com.launchdarkly.android.Util.isInternetConnected;
 
 /**
@@ -36,20 +38,25 @@ import static com.launchdarkly.android.Util.isInternetConnected;
  * The main entry point is the {@link #init(Application, LDConfig, LDUser)} method.
  */
 public class LDClient implements LDClientInterface, Closeable {
-    private static final String TAG = "LaunchDarkly";
 
     private static final String INSTANCE_ID_KEY = "instanceId";
     // Upon client init will get set to a Unique id per installation used when creating anonymous users
     private static String instanceId = "UNKNOWN_ANDROID";
     private static LDClient instance = null;
 
+    private static final long MAX_RETRY_TIME_MS = 3600000; // 1 hour
+    private static final long RETRY_TIME_MS = 1000; // 1 second
+
+    private final WeakReference<Application> application;
     private final LDConfig config;
     private final UserManager userManager;
     private final EventProcessor eventProcessor;
     private final UpdateProcessor updateProcessor;
     private final FeatureFlagFetcher fetcher;
+    private final Throttler throttler;
 
     private volatile boolean isOffline = false;
+    private volatile boolean isAppForegrounded = true;
 
     /**
      * Initializes the singleton instance. The result is a {@link Future} which
@@ -84,7 +91,7 @@ public class LDClient implements LDClientInterface, Closeable {
         SettableFuture<LDClient> settableFuture = SettableFuture.create();
 
         if (instance != null) {
-            Log.w(TAG, "LDClient.init() was called more than once! returning existing instance.");
+            Timber.w( "LDClient.init() was called more than once! returning existing instance.");
             settableFuture.set(instance);
             return settableFuture;
         }
@@ -134,14 +141,14 @@ public class LDClient implements LDClientInterface, Closeable {
      * @return
      */
     public static synchronized LDClient init(Application application, LDConfig config, LDUser user, int startWaitSeconds) {
-        Log.i(TAG, "Initializing Client and waiting up to " + startWaitSeconds + " for initialization to complete");
+        Timber.i("Initializing Client and waiting up to " + startWaitSeconds + " for initialization to complete");
         Future<LDClient> initFuture = init(application, config, user);
         try {
             return initFuture.get(startWaitSeconds, TimeUnit.SECONDS);
         } catch (InterruptedException | ExecutionException e) {
-            Log.e(TAG, "Exception during Client initialization", e);
+            Timber.e(e, "Exception during Client initialization");
         } catch (TimeoutException e) {
-            Log.w(TAG, "Client did not successfully initialize within " + startWaitSeconds + " seconds. " +
+            Timber.w("Client did not successfully initialize within " + startWaitSeconds + " seconds. " +
                     "It could be taking longer than expected to start up");
         }
         return instance;
@@ -153,7 +160,7 @@ public class LDClient implements LDClientInterface, Closeable {
      */
     public static LDClient get() throws LaunchDarklyException {
         if (instance == null) {
-            Log.e(TAG, "LDClient.get() was called before init()!");
+            Timber.e("LDClient.get() was called before init()!");
             throw new LaunchDarklyException("LDClient.get() was called before init()!");
         }
         return instance;
@@ -161,22 +168,23 @@ public class LDClient implements LDClientInterface, Closeable {
 
     @VisibleForTesting
     protected LDClient(final Application application, @NonNull final LDConfig config) {
-        Log.i(TAG, "Creating LaunchDarkly client. Version: " + BuildConfig.VERSION_NAME);
+        Timber.i("Creating LaunchDarkly client. Version: %s", BuildConfig.VERSION_NAME);
         this.config = config;
         this.isOffline = config.isOffline();
+        this.application = new WeakReference<>(application);
 
         SharedPreferences instanceIdSharedPrefs = application.getSharedPreferences(LDConfig.SHARED_PREFS_BASE_KEY + "id", Context.MODE_PRIVATE);
 
         if (!instanceIdSharedPrefs.contains(INSTANCE_ID_KEY)) {
             String uuid = UUID.randomUUID().toString();
-            Log.i(TAG, "Did not find existing instance id. Saving a new one");
+            Timber.i("Did not find existing instance id. Saving a new one");
             SharedPreferences.Editor editor = instanceIdSharedPrefs.edit();
             editor.putString(INSTANCE_ID_KEY, uuid);
             editor.apply();
         }
 
         instanceId = instanceIdSharedPrefs.getString(INSTANCE_ID_KEY, instanceId);
-        Log.i(TAG, "Using instance id: " + instanceId);
+        Timber.i("Using instance id: " + instanceId);
 
         this.fetcher = HttpFeatureFlagFetcher.init(application, config);
         this.userManager = UserManager.init(application, fetcher);
@@ -185,7 +193,8 @@ public class LDClient implements LDClientInterface, Closeable {
             @Override
             public void onBecameForeground() {
                 PollingUpdater.stop(application);
-                if (!isOffline() && isInternetConnected(application)) {
+                isAppForegrounded = true;
+                if (isInternetConnected(application)) {
                     startForegroundUpdating();
                 }
             }
@@ -193,9 +202,8 @@ public class LDClient implements LDClientInterface, Closeable {
             @Override
             public void onBecameBackground() {
                 stopForegroundUpdating();
-                if (!config.isDisableBackgroundPolling() && !isOffline() && isInternetConnected(application)) {
-                    PollingUpdater.startBackgroundPolling(application);
-                }
+                isAppForegrounded = false;
+                startBackgroundPolling();
             }
         };
         foreground.addListener(foregroundListener);
@@ -203,10 +211,17 @@ public class LDClient implements LDClientInterface, Closeable {
         if (config.isStream()) {
             this.updateProcessor = new StreamUpdateProcessor(config, userManager);
         } else {
-            Log.i(TAG, "Streaming is disabled. Starting LaunchDarkly Client in polling mode");
+            Timber.i("Streaming is disabled. Starting LaunchDarkly Client in polling mode");
             this.updateProcessor = new PollingUpdateProcessor(application, userManager, config);
         }
         eventProcessor = new EventProcessor(application, config);
+
+        throttler = new Throttler(new Runnable() {
+            @Override
+            public void run() {
+                setOnlineStatus();
+            }
+        }, RETRY_TIME_MS, MAX_RETRY_TIME_MS);
     }
 
     /**
@@ -244,7 +259,7 @@ public class LDClient implements LDClientInterface, Closeable {
         }
 
         if (user.getKey() == null) {
-            Log.w(TAG, "identify called with null user or null user key!");
+            Timber.w("identify called with null user or null user key!");
         }
 
         Future<Void> doneFuture;
@@ -296,11 +311,11 @@ public class LDClient implements LDClientInterface, Closeable {
         try {
             result = userManager.getCurrentUserSharedPrefs().getBoolean(flagKey, fallback);
         } catch (ClassCastException cce) {
-            Log.e(TAG, "Attempted to get boolean flag that exists as another type for key: "
-                    + flagKey + " Returning fallback: " + fallback, cce);
+            Timber.e(cce, "Attempted to get boolean flag that exists as another type for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         } catch (NullPointerException npe) {
-            Log.e(TAG, "Attempted to get boolean flag with a default null value for key: "
-                    + flagKey + " Returning fallback: " + fallback, npe);
+            Timber.e(npe, "Attempted to get boolean flag with a default null value for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         }
         if (result == null && fallback == null) {
             sendFlagRequestEvent(flagKey, JsonNull.INSTANCE, JsonNull.INSTANCE);
@@ -311,7 +326,7 @@ public class LDClient implements LDClientInterface, Closeable {
         } else {
             sendFlagRequestEvent(flagKey, new JsonPrimitive(result), new JsonPrimitive(fallback));
         }
-        Log.d(TAG, "boolVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
+        Timber.d("boolVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
         return result;
     }
 
@@ -333,11 +348,11 @@ public class LDClient implements LDClientInterface, Closeable {
         try {
             result = (int) userManager.getCurrentUserSharedPrefs().getFloat(flagKey, fallback);
         } catch (ClassCastException cce) {
-            Log.e(TAG, "Attempted to get integer flag that exists as another type for key: "
-                    + flagKey + " Returning fallback: " + fallback, cce);
+            Timber.e(cce, "Attempted to get integer flag that exists as another type for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         } catch (NullPointerException npe) {
-            Log.e(TAG, "Attempted to get integer flag with a default null value for key: "
-                    + flagKey + " Returning fallback: " + fallback, npe);
+            Timber.e(npe, "Attempted to get integer flag with a default null value for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         }
         if (result == null && fallback == null) {
             sendFlagRequestEvent(flagKey, JsonNull.INSTANCE, JsonNull.INSTANCE);
@@ -348,7 +363,7 @@ public class LDClient implements LDClientInterface, Closeable {
         } else {
             sendFlagRequestEvent(flagKey, new JsonPrimitive(result), new JsonPrimitive(fallback));
         }
-        Log.d(TAG, "intVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
+        Timber.d("intVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
         return result;
     }
 
@@ -370,11 +385,11 @@ public class LDClient implements LDClientInterface, Closeable {
         try {
             result = userManager.getCurrentUserSharedPrefs().getFloat(flagKey, fallback);
         } catch (ClassCastException cce) {
-            Log.e(TAG, "Attempted to get float flag that exists as another type for key: "
-                    + flagKey + " Returning fallback: " + fallback, cce);
+            Timber.e(cce, "Attempted to get float flag that exists as another type for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         } catch (NullPointerException npe) {
-            Log.e(TAG, "Attempted to get float flag with a default null value for key: "
-                    + flagKey + " Returning fallback: " + fallback, npe);
+            Timber.e(npe, "Attempted to get float flag with a default null value for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         }
         if (result == null && fallback == null) {
             sendFlagRequestEvent(flagKey, JsonNull.INSTANCE, JsonNull.INSTANCE);
@@ -385,7 +400,7 @@ public class LDClient implements LDClientInterface, Closeable {
         } else {
             sendFlagRequestEvent(flagKey, new JsonPrimitive(result), new JsonPrimitive(fallback));
         }
-        Log.d(TAG, "floatVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
+        Timber.d("floatVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
         return result;
     }
 
@@ -407,11 +422,11 @@ public class LDClient implements LDClientInterface, Closeable {
         try {
             result = userManager.getCurrentUserSharedPrefs().getString(flagKey, fallback);
         } catch (ClassCastException cce) {
-            Log.e(TAG, "Attempted to get string flag that exists as another type for key: "
-                    + flagKey + " Returning fallback: " + fallback, cce);
+            Timber.e(cce, "Attempted to get string flag that exists as another type for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         } catch (NullPointerException npe) {
-            Log.e(TAG, "Attempted to get string flag with a default null value for key: "
-                    + flagKey + " Returning fallback: " + fallback, npe);
+            Timber.e(npe, "Attempted to get string flag with a default null value for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         }
         if (result == null && fallback == null) {
             sendFlagRequestEvent(flagKey, JsonNull.INSTANCE, JsonNull.INSTANCE);
@@ -422,7 +437,7 @@ public class LDClient implements LDClientInterface, Closeable {
         } else {
             sendFlagRequestEvent(flagKey, new JsonPrimitive(result), new JsonPrimitive(fallback));
         }
-        Log.d(TAG, "stringVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
+        Timber.d("stringVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
         return result;
     }
 
@@ -447,17 +462,17 @@ public class LDClient implements LDClientInterface, Closeable {
                 result = new JsonParser().parse(stringResult);
             }
         } catch (ClassCastException cce) {
-            Log.e(TAG, "Attempted to get json (string) flag that exists as another type for key: "
-                    + flagKey + " Returning fallback: " + fallback, cce);
+            Timber.e(cce, "Attempted to get json (string) flag that exists as another type for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         } catch (NullPointerException npe) {
-            Log.e(TAG, "Attempted to get json (string flag with a default null value for key: "
-                    + flagKey + " Returning fallback: " + fallback, npe);
+            Timber.e(npe, "Attempted to get json (string flag with a default null value for key: "
+                    + flagKey + " Returning fallback: " + fallback);
         } catch (JsonSyntaxException jse) {
-            Log.e(TAG, "Attempted to get json (string flag that exists as another type for key: " +
-                    flagKey + " Returning fallback: " + fallback, jse);
+            Timber.e(jse, "Attempted to get json (string flag that exists as another type for key: " +
+                    flagKey + " Returning fallback: " + fallback);
         }
         sendFlagRequestEvent(flagKey, result, fallback);
-        Log.d(TAG, "jsonVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
+        Timber.d("jsonVariation: returning variation: " + result + " flagKey: " + flagKey + " user key: " + userManager.getCurrentUser().getKeyAsString());
         return result;
     }
 
@@ -490,21 +505,47 @@ public class LDClient implements LDClientInterface, Closeable {
         return isOffline;
     }
 
+    /**
+     * Shuts down any network connections maintained by the client and puts the client in offline
+     * mode, preventing the client from opening new network connections until
+     * <code>setOnline()</code> is called.
+     * <p/>
+     * Note: The client automatically monitors the device's network connectivity and app foreground
+     * status, so calling <code>setOffline()</code> or <code>setOnline()</code> is normally
+     * unnecessary in most situations.
+     */
     @Override
     public synchronized void setOffline() {
-        Log.d(TAG, "Setting isOffline = true");
+        Timber.d("Setting isOffline = true");
+        throttler.cancel();
         isOffline = true;
         fetcher.setOffline();
         stopForegroundUpdating();
         eventProcessor.stop();
     }
 
+    /**
+     * Restores network connectivity for the client, if the client was previously in offline mode.
+     * This operation may be throttled if it is called too frequently.
+     * <p/>
+     * Note: The client automatically monitors the device's network connectivity and app foreground
+     * status, so calling <code>setOffline()</code> or <code>setOnline()</code> is normally
+     * unnecessary in most situations.
+     */
     @Override
     public synchronized void setOnline() {
-        Log.d(TAG, "Setting isOffline = false");
-        this.isOffline = false;
+        throttler.attemptRun();
+    }
+
+    private void setOnlineStatus() {
+        Timber.d("Setting isOffline = false");
+        isOffline = false;
         fetcher.setOnline();
-        startForegroundUpdating();
+        if (isAppForegrounded) {
+            startForegroundUpdating();
+        } else {
+            startBackgroundPolling();
+        }
         eventProcessor.start();
     }
 
@@ -550,6 +591,13 @@ public class LDClient implements LDClientInterface, Closeable {
         }
     }
 
+    void startBackgroundPolling() {
+        Application application = this.application.get();
+        if (application != null && !config.isDisableBackgroundPolling() && !isOffline() && isInternetConnected(application)) {
+            PollingUpdater.startBackgroundPolling(application);
+        }
+    }
+
     private void sendFlagRequestEvent(String flagKey, JsonElement value, JsonElement fallback) {
         sendEvent(new FeatureRequestEvent(flagKey, userManager.getCurrentUser(), value, fallback));
     }
@@ -558,7 +606,7 @@ public class LDClient implements LDClientInterface, Closeable {
         if (!isOffline()) {
             boolean processed = eventProcessor.sendEvent(event);
             if (!processed) {
-                Log.w(TAG, "Exceeded event queue capacity. Increase capacity to avoid dropping events.");
+                Timber.w("Exceeded event queue capacity. Increase capacity to avoid dropping events.");
             }
         }
     }